requirements.txt
pillow>=10.0.0
opencv-python>=4.9.0.80
numpy>=1.24.0,<2.1
rembg>=2.0.50
torch>=2.2.0
